Chrisjurich/feature new mutation submodule

+++ enzy_htp/preparation/pdb_prepper.py
@@ -7,6 +7,7 @@ Date: 2022-03-12
 """
 import shutil
 import numpy as np
+from pathlib import Path
 
 import pdb2pqr
 from typing import Set, Union, List, Tuple
@@ -28,28 +29,18 @@ from enzy_htp.structure import (
 )
 
 from .pdb_line import PDBLine, read_pdb_lines
-
-
-from .mutate import (
-    Mutation,
-    mutation_to_str,
-    decode_mutaflags,
-    get_all_combinations,
-    get_all_combinations,
-)
-
 from enzy_htp.chemical import convert_to_three_letter, get_element_aliases
 
 
 class PDBPrepper:
     """Class that handles initial preparation of a specified Structure().
 
-	Attributes:
+    Attributes:
         current_path_ : The path to the most recently saved version of the Structure().
-	    work_dir : The scratch directory where temporary files are saved. The current time in the format "YYYY_MM_DD" is used if the keyword is not specified.
-		all_paths : A list of all paths created by the PDBPrepper() object. Stored in reverse order of age.
+        work_dir : The scratch directory where temporary files are saved. The current time in the format "YYYY_MM_DD" is used if the keyword is not specified.
+        all_paths : A list of all paths created by the PDBPrepper() object. Stored in reverse order of age.
 
-	"""
+    """
 
     def __init__(self, pdb_name, **kwargs):
         """Inits PDBPrepper with a pdb file and optionally a work directory to place temporary files."""
@@ -61,10 +52,11 @@ class PDBPrepper:
             "work_dir", f"{fs.get_current_time()}_{self.base_pdb_name}"
         )
         fs.safe_mkdir(self.work_dir)
-        shutil.copy(self.pdb_path, f"{self.work_dir}/{self.base_pdb_name}.pdb")
+        if not Path(f"{self.work_dir}/{self.base_pdb_name}.pdb").exists(): #TODO(CJ): Do we need a safe_cpy function?
+            shutil.copy(self.pdb_path, f"{self.work_dir}/{self.base_pdb_name}.pdb")
         self.path_name = f"{self.work_dir}/{self.base_pdb_name}.pdb"
         self.pqr_path = str()
-        self.mutations = []
+        # self.mutations = []
         self.current_path_ = self.path_name
         self.all_paths = [self.current_path_]
 
@@ -277,7 +269,13 @@ class PDBPrepper:
                         # Deal with missing residue, fill with "NAN"
                         missing_length = pdb_l.resi_id - last_resi_index - 1
                         if missing_length > 0:
-                            Chain_sequence = Chain_sequence + ["NAN",] * missing_length
+                            Chain_sequence = (
+                                Chain_sequence
+                                + [
+                                    "NAN",
+                                ]
+                                * missing_length
+                            )
 
                         # Store the new resi
                         Chain_sequence.append(pdb_l.resi_name)
@@ -408,10 +406,14 @@ class PDBPrepper:
 
         pass
 
-    def get_protonation(self, ph: float = 7.0, ffout: str = "AMBER",) -> str:
+    def get_protonation(
+        self,
+        ph: float = 7.0,
+        ffout: str = "AMBER",
+    ) -> str:
         """Uses PDB2PQR to get the protonation state for a PDB file, saving the protonated .pdb file.
-		Separately runs PDB2PQR on the ligand and metal and combines the output. Returns path to the protonated
-		PDB file."""
+        Separately runs PDB2PQR on the ligand and metal and combines the output. Returns path to the protonated
+        PDB file."""
         # input of PDB2PQR
         check_valid_ph(ph)
         self.pqr_path = f"{fs.remove_ext(self.current_path_)}.pqr.pdb"
@@ -431,141 +433,7 @@ class PDBPrepper:
 
         return self.current_path_
 
-    """
-    ========
-    Mutation
-    ========
-    """
-
-    def apply_mutations(self) -> str:
-        """
-        Apply mutations using tleap. Save mutated structure PDB in self.path
-        ------------------------------
-        Use MutaFlag in self.MutaFlags
-        Grammer (from Add_MutaFlag):
-        X : Original residue name. Leave X if unknow. 
-            Only used for build filenames. **Do not affect any calculation.**
-        A : Chain index. Determine by 'TER' marks in the PDB file. (Do not consider chain_indexs in the original file.)
-        11: Residue index. Strictly correponding residue indexes in the original file. (NO sort applied)
-        Y : Target residue name.  
-
-        **WARNING** if there are multiple mutations on the same index, only the first one will be used.
-        """
-
-        # Prepare a label for the filename
-        tot_Flag_name = "_".join(list(map(mutaflag_to_str, self.mutations)))
-        # Operate the PDB
-        out_PDB_path1 = (
-            self.work_dir + "/" + self.base_pdb_name + tot_Flag_name + "_tmp.pdb"
-        )
-        out_PDB_path2 = self.path_name + tot_Flag_name + ".pdb"
-        chain_count = 1
-        pdb_lines = read_pdb_lines(self.path_name)
-        mask = [True] * len(pdb_lines)
-        for pdb_l in pdb_lines:
-            if pdb_l.is_TER():
-                chain_count += 1
-            match = 0
-            # only match in the dataline and keep all non data lines
-            if not pdb_l.is_ATOM():
-                continue
-            for mf in self.mutations:
-                # Test for every Flag for every lines
-
-                if (
-                    chr(64 + chain_count) == mf.chain_index
-                    and pdb_l.resi_id == mf.residue_index
-                ):
-                    # do not write old line if match a MutaFlag
-                    match = 1
-                    # Keep OldAtoms of targeted old residue
-                    target_residue = mf.target_residue
-                    # fix for mutations of Gly & Pro
-                    old_atoms = {
-                        "G": ["N", "H", "CA", "C", "O"],
-                        "P": ["N", "CA", "HA", "CB", "C", "O"],
-                    }.get(mf.target_residue, ["N", "H", "CA", "HA", "CB", "C", "O"])
-
-                    line = pdb_l.line
-                    for oa in old_atoms:
-                        if oa == pdb_l.atom_name:
-                            pdb_l.line = f"{line[:17]}{convert_to_three_letter(mf.target_residue)}{line[20:]}"
-
-        write_lines(out_PDB_path1, list(map(lambda pl: pl.line, pdb_lines)))
-        leapin_path = f"{self.work_dir}/leap_P2PwL.in"
-        leap_lines = [
-            "source leaprc.protein.ff14SB",
-            f"a = loadpdb {out_PDB_path1}",
-            f"savepdb a {out_PDB_path2}",
-            "quit",
-        ]
-        write_lines(leapin_path, leap_lines)
-        em.run_command(
-            "tleap", [f"-s -f {leapin_path} > {self.work_dir}/leap_P2PwL.out"]
-        )
-        safe_rm("leap.log")
-
-    def generate_mutations(
-        self,
-        n: int,
-        muta_flags: Union[str, List[str]] = None,
-        restrictions: List[Union[int, Tuple[int, int]]] = None,
-        random_state: int = 100,
-    ) -> None:
-        """"""
-        existing = set()
-        temp = []
-        if muta_flags:
-            decoded: List[MutaFlag] = decode_mutaflags(muta_flags)
-            for mf in decoded:
-                key = (mf.chain_index, mf.residue_index)
-                if key in existing:
-                    _LOGGER.warn(f"Multiple mutations supplied at location {key}")
-                else:
-                    existing.add(key)
-                    temp.append(mf)
-
-        muta_flags = temp
-
-        if muta_flags and len(muta_flags) >= n:
-            _LOGGER.warning(
-                f"Supplied mutation flags meet or exceed the number of desired mutations"
-            )
-            self.mutations = muta_flags
-            return
-        struct: Structure = structure_from_pdb(self.current_path())
-        curr_state = struct.residue_state()
-        candidates: List[MutaFlag] = get_all_combinations(
-            curr_state, restrictions=restrictions
-        )
-
-        np.random.seed(random_state)
-        np.random.shuffle(candidates)
-
-        while len(muta_flags) < n and len(candidates):
-            curr = candidates.pop()
-            key = (curr.chain_index, curr.residue_index)
-            if key in existing:
-                continue
-            else:
-                muta_flags.append(curr)
-                existing.add(key)
-
-        if len(muta_flags) != n:
-            _LOGGER.warning(
-                f"Unable to generate enough mutations. Missing {n-len(muta_flags)}"
-            )
-
-        self.mutations = muta_flags
-        assert len(self.mutations) == len(set(self.mutations))
-
-    def _build_MutaName(self, Flag):
-        """
-        Take a MutaFlag Tuple and return a str of name
-        """
-        return Flag[0] + Flag[1] + Flag[2] + Flag[3]
-
-    def rm_allH(self, ff="Amber", ligand=False):
+    def rm_allH(self, ff:str="Amber", ligand:bool=False):
         # TODO: CJ: should this get called by mutation method?
         """
         remove wrong hydrogens added by leap after mutation. (In the case that the input file was a H-less one from crystal.)
@@ -575,8 +443,8 @@ class PDBPrepper:
         1 - remove all Hs base on the nomenclature. (start with H and not in the non_H_list)
         """
         # out path
-        o_path = self.path_name + "_rmH.pdb"
-        pdb_lines: List[PDBLine] = read_pdb_lines(self.curr_path)
+        o_path = self.path_name.replace('.pdb','') + "_rmH.pdb"
+        pdb_lines: List[PDBLine] = read_pdb_lines(self.current_path_)
         mask = [True] * len(pdb_lines)
         # crude judgement of H including customized H
         if ligand:
@@ -594,9 +462,10 @@ class PDBPrepper:
             for idx, pl in enumerate(pdb_lines):
                 if pl.atom_name in H_aliases and pl.is_residue_line():
                     mask[idx] = False
+                #print(pl.line, mask[idx])
 
         pdb_lines = np.array(pdb_lines)[mask]
-        write_lines(o_path, list(map(str, pdb_lines)))
+        fs.write_lines(o_path, list(map(str, pdb_lines)))
 
     """
     ========
@@ -645,87 +514,6 @@ class PDBPrepper:
     QM Cluster
     ========    
     """
-
-    def PDB2QMCluster(
-        self,
-        atom_mask,
-        spin=1,
-        o_dir="",
-        tag="",
-        QM="g16",
-        g_route=None,
-        ifchk=0,
-        val_fix="internal",
-    ):
-        """
-        Build & Run QM cluster input from self.mdcrd with selected atoms according to atom_mask
-        ---------
-        spin: specific spin state for the qm cluster. (default: 1)
-        QM: QM engine (default: Gaussian)
-            g_route: Gaussian route line
-            ifchk: if save chk file of a gaussian job. (default: 0)
-        val_fix: fix free valance of truncated strutures
-            = internal: add H to where the original connecting atom is. 
-                        special fix for classical case:
-                        "sele by residue" (cut N-C) -- adjust dihedral for added H on N.
-            = openbabel TODO
-        ---data---
-        self.frames
-        self.qm_cluster_map (PDB atom id -> QM atom id)
-        """
-        # make folder
-        if o_dir == "":
-            o_dir = self.dir + "/QM_cluster" + tag
-        mkdir(o_dir)
-        # update stru
-        self.get_stru()
-        # get sele
-        if val_fix == "internal":
-            sele_lines, sele_map = self.stru.get_sele_list(
-                atom_mask, fix_end="H", prepi_path=self.prepi_path
-            )
-        else:
-            sele_lines, sele_map = self.stru.get_sele_list(atom_mask, fix_end=None)
-        self.qm_cluster_map = sele_map
-        # get chrgspin
-        chrgspin = self._get_qmcluster_chrgspin(sele_lines, spin=spin)
-        if Config.debug >= 1:
-            print("Charge: " + str(chrgspin[0]) + " Spin: " + str(chrgspin[1]))
-
-        # make inp files
-        frames = Frame.fromMDCrd(self.mdcrd)
-        self.frames = frames
-        if QM in ["g16", "g09"]:
-            gjf_paths = []
-            if Config.debug >= 1:
-                print("Writing QMcluster gjfs.")
-            for i, frame in enumerate(frames):
-                gjf_path = o_dir + "/qm_cluster_" + str(i) + ".gjf"
-                frame.write_sele_lines(
-                    sele_lines,
-                    out_path=gjf_path,
-                    g_route=g_route,
-                    chrgspin=chrgspin,
-                    ifchk=ifchk,
-                )
-                gjf_paths.append(gjf_path)
-            # Run inp files
-            qm_cluster_out_paths = PDB.Run_QM(gjf_paths, prog=QM)
-            # get chk files if ifchk
-            if ifchk:
-                qm_cluster_chk_paths = []
-                for gjf in gjf_paths:
-                    chk_path = gjf[:-3] + "chk"
-                    qm_cluster_chk_paths.append(chk_path)
-        if QM == "ORCA":
-            pass
-
-        self.qm_cluster_out = qm_cluster_out_paths
-        if ifchk:
-            self.qm_cluster_chk = qm_cluster_chk_paths
-            return self.qm_cluster_out, self.qm_cluster_chk
-
-        return self.qm_cluster_out
 
     def _get_qmcluster_chrgspin(self, sele, spin=1):
         """
